requirements.txt
requests
google-cloud-storage
shodan